requirements.txt
aiohttp==3.7.4.post0
alembic==1.8.0
altgraph==0.17.2
asgiref==3.5.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
cycler==0.11.0
DAWG-Python==0.7.2
dialogic==0.3.20
discord.py==1.7.3
Django==4.0.5
django-autoslug==1.9.8
dnspython==2.2.1
docopt==0.6.2
email-validator==1.2.1
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.33.3
future==0.18.2
fuzzywuzzy==0.18.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
huggingface-hub==0.5.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.3
lxml==4.9.0
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
numpy==1.22.3
openpyxl==3.0.9
outcome==1.2.0
packaging==21.3
pandas==1.4.2
parsel==1.6.0
pefile==2022.5.30
Pillow==9.1.1
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pymessenger==0.0.7.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32-ctypes==0.2.0
PyYAML==6.0
queuelib==1.6.2
regex==2022.4.24
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
Scrapy==2.6.1
selenium==4.3.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
sqlparse==0.4.2
textdistance==4.2.2
threadpoolctl==3.1.0
tldextract==3.3.0
tokenizers==0.12.1
tqdm==4.64.0
transliterate==1.10.2
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
w3lib==1.22.0
Werkzeug==2.1.2
wsproto==1.1.0
WTForms==3.0.1
yarl==1.7.2
zipp==3.8.0
zope.interface==5.4.0
